requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
pluggy==1.5.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.2.1
SQLAlchemy==2.0.29
tomli==2.0.1
typing_extensions==4.11.0
yarl==1.9.4
python-dotenv~=1.0.1